+++ rotkehlchen/chain/base/modules/hop/decoder.py
@@ -2,7 +2,7 @@ from typing import TYPE_CHECKING
 
 from rotkehlchen.chain.evm.decoding.hop.decoder import HopCommonDecoder
 
-from .constants import BRIDGES
+from .constants import BRIDGES, REWARD_CONTRACTS
 
 if TYPE_CHECKING:
     from rotkehlchen.chain.base.node_inquirer import BaseInquirer
@@ -23,4 +23,5 @@ class HopDecoder(HopCommonDecoder):
             base_tools=base_tools,
             msg_aggregator=msg_aggregator,
             bridges=BRIDGES,
+            reward_contracts=REWARD_CONTRACTS,
         )